Actual Content: The commit message for the changes ade in the dockerfile that was modified.
Actualizied Content: Made changes to the dockerfile to include the latest version of
Node.js v18 (actual LTS). Also updated the Dockerfile to include the latest
version of npm v8.0.0. Updated the Dockerfile to include the latest version of
yarn v1.22.17. Updated the Dockerfile to include a multi-stage build. The first
stage is the builder stage which installs the dependencies and builds the
application. The second stage is the production stage which copies the built
application from the builder stage and runs the application.

Multi Staging Tools:
- Node.js v18
- npm v8.0.0
- yarn v1.22.17

Multi-Staged Dockerfile:
- FROM node:18 AS builder
- WORKDIR /opt/sql-language-server
- COPY . .
- RUN yarn install && yarn build